Extracted Stage (Window) creation from the main method.
Still odd though, as the main method in JavaFX comes from implementing Application in the main java file. Which provides a stage itself.

Maybe a better solution would be to use the stage provided by Application, and provide it with a master Scene (which will always be the root view in the window).